+++ sdks/csharp/examples/BasicUsage/obj/Debug/net8.0/BasicUsage.AssemblyInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("BasicUsage")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+9400a0de5097e991dc4fee7810e52ad1b9feab82")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+b652918fff1a8353c7b89f9c72bca62765a1baaf")]
 [assembly: System.Reflection.AssemblyProductAttribute("BasicUsage")]
 [assembly: System.Reflection.AssemblyTitleAttribute("BasicUsage")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]